+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3890,7 +3890,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01M2NEZRM1XWDP7B1GQKY7E8XN"),
+  id: z.string().default("01M2PKNB414QS7W5JR0BVJC100"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),